# packages/client/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rhizome rNet API",
    "version": "0.1.0",
    "description": "The rNet-compatible API exposed by the Rhizome store."
  },
  "servers": [{ "url": "/", "description": "Current Rhizome store" }],
  "paths": {
    "/rnet/v0/vibes": {
      "get": {
        "operationId": "listVibes",
        "tags": ["Vibes"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["vibes"],
                  "properties": {
                    "vibes": { "type": "array", "items": { "$ref": "#/components/schemas/Vibe" } }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      },
      "post": {
        "operationId": "createVibe",
        "tags": ["Vibes"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 256 },
                  "pull": {
                    "type": "object",
                    "description": "How the Vibe acquires new objects.",
                    "required": ["enabled"],
                    "properties": {
                      "enabled": { "type": "boolean" },
                      "sources": {
                        "type": "array",
                        "description": "Source identifiers, implementation-scoped (e.g. \"skill:ofx@^0.3\").",
                        "items": { "type": "string", "minLength": 1 }
                      },
                      "policy": { "enum": ["append_new", "replace", "suggest_only"] },
                      "last_pulled_at": { "type": "string", "format": "date-time" }
                    },
                    "additionalProperties": false
                  },
                  "grants": { "type": "array", "items": { "$ref": "#/components/schemas/Grant" } }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Vibe" } } }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
    "/rnet/v0/objects": {
      "post": {
        "operationId": "createMediaObjects",
        "tags": ["Media Objects"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user_or_client",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["metadata"],
                "properties": {
                  "metadata": {
                    "type": "string",
                    "contentMediaType": "application/json",
                    "contentSchema": {
                      "type": "object",
                      "required": ["objects"],
                      "properties": {
                        "vibe": {
                          "type": "string",
                          "pattern": "^rnet://vibe/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                        },
                        "objects": {
                          "type": "array",
                          "minItems": 1,
                          "items": {
                            "type": "object",
                            "properties": {
                              "elements": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string",
                                      "pattern": "^rnet://element/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                                    },
                                    {
                                      "type": "object",
                                      "required": ["upload", "kind", "mime"],
                                      "properties": {
                                        "upload": { "type": "string", "minLength": 1 },
                                        "kind": {
                                          "description": "Consumption strategy. Closed set — new kinds only by protocol revision.",
                                          "enum": ["text", "image", "audio", "video", "document"]
                                        },
                                        "mime": {
                                          "type": "string",
                                          "description": "IANA media type. mime is truth; kind is the consumption hint.",
                                          "pattern": "^[a-z]+/[a-zA-Z0-9][a-zA-Z0-9!#$&^_.+-]*$"
                                        }
                                      },
                                      "additionalProperties": false
                                    }
                                  ]
                                }
                              }
                            }
                          }
                        }
                      },
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["mediaObjects"],
                  "properties": {
                    "mediaObjects": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MediaObject" }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "415": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
    "/rnet/v0/elements": {
      "post": {
        "operationId": "createMediaElement",
        "tags": ["Media Elements"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user",
        "requestBody": {
          "required": true,
          "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MediaElement" } }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/elements/{id}": {
      "get": {
        "operationId": "getMediaElement",
        "tags": ["Media Elements"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MediaElement" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/elements/{id}/bytes": {
      "get": {
        "operationId": "getMediaElementBytes",
        "tags": ["Media Elements"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary content",
            "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/origins": {
      "post": {
        "operationId": "createOriginArtifact",
        "tags": ["Origin Artifacts"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user",
        "requestBody": {
          "required": true,
          "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OriginArtifact" } }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "415": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "BearerAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
      "Grant": {
        "title": "Grant",
        "description": "The read-edge conformance unit: {subject, scope[]} attached to a Vibe. Scope semantics are store-enforced, always — never client-honor-system. Grants are owner-set; delegation is not specified. Stores MUST deny grants whose subject namespace they do not understand.",
        "type": "object",
        "required": ["subject", "scope"],
        "properties": {
          "subject": {
            "type": "string",
            "description": "Opaque string with a namespace prefix: id:{rnet-id} (a user), client:{name} (a registered application), public (anyone), or {x-namespace}:{...} (extension subject types).",
            "pattern": "^(id:rnet://id/[A-Za-z0-9._~-]+|client:[a-z0-9][a-z0-9._-]*|public|x-[a-z0-9-]+:.+)$"
          },
          "scope": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "enum": ["read", "write:user", "write:objects", "write:inferred", "push", "pull"]
            },
            "description": "read: fetch Vibe/objects/elements — never origins, which are owner-only. write:user: mutate user blocks only. write:objects: create authored objects. write:inferred: write inferred entries directly, under a subject-namespaced task key. push: invoke push. pull: invoke a pull on already-configured sources. Grants are set by the owner alone — no scope delegates the ability to grant."
          }
        },
        "additionalProperties": false
      },
...
      "MediaObject": {
        "title": "MediaObject",
        "description": "The core unit of meaning: properties plus zero or more MediaElements, carrying three property blocks with different mutation rights. Fields are values, elements are files — if you would query on it, it is a field; if you would open it, it is an element. Pure meaning-objects legitimately carry zero elements.",
        "type": "object",
        "required": ["rnet_schema", "uri", "owner", "type", "elements", "source"],
        "properties": {
          "rnet_schema": {
            "const": "0.1",
            "description": "Protocol version this document conforms to."
          },
          "uri": {
            "type": "string",
            "description": "UUIDv7-identified record.",
            "pattern": "^rnet://object/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "owner": {
            "type": "string",
            "description": "Immutable owner identity assigned by the store at creation. Ownership governs administration, not delegated access.",
            "pattern": "^rnet://id/[A-Za-z0-9._~-]+$"
          },
          "type": {
            "type": "string",
            "description": "Open vocabulary. The registered core vocabulary currently includes transaction and track; unregistered types such as post, photo, note, contact, event, book, article, and receipt remain legal.",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[a-z][a-z0-9_.-]*$"
          },
          "elements": {
            "type": "array",
            "description": "Ordered MediaElement URIs. MAY be empty.",
            "items": {
              "type": "string",
              "pattern": "^rnet://element/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          "keys": {
            "type": "object",
            "description": "External global identifiers for cross-service joins: isrc, isbn, fitid, url, ean, etc.",
            "additionalProperties": { "type": "string", "maxLength": 512 }
          },
          "source": {
            "type": "object",
            "description": "Written by ingestion runtimes only. Immutable after ingest — re-ingestion creates a new revision, never edits in place.",
            "required": ["ingest", "origins", "properties"],
            "properties": {
              "ingest": { "$ref": "#/components/schemas/IngestRecord" },
              "origins": {
                "type": "array",
                "minItems": 1,
                "description": "What this object was derived from — at least one, always. Either an origin artifact (rnet://origin/{uuid}) for ingested objects, or a client (rnet://client/{uuid}) for objects authored directly in an application. A filename string is not provenance.",
                "items": {
                  "type": "string",
                  "pattern": "^rnet://(origin|client)/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "retrieved_at": { "type": "string", "format": "date-time" },
              "properties": {
                "type": "object",
                "description": "Facts the source format defines. Validated against the registered type vocabulary when one exists."
              }
            },
            "additionalProperties": false
          },
          "user": {
            "type": "object",
            "description": "Written by the owner via clients holding write:user. Freely mutable; revision-protected.",
            "required": ["properties"],
            "properties": {
              "properties": { "type": "object" },
              "updated_at": { "type": "string", "format": "date-time" }
            },
            "additionalProperties": false
          },
          "inferred": {
            "type": "object",
            "description": "A map keyed by writer and task: every key is {writer}:{task}. Memory scoped to this record — some entries are task output recomputed from source, others accumulated from corrections and agent observation and cannot be re-derived. A re-run replaces only its own key, and never a durable entry. Consumers MUST treat entries as advisory.",
            "propertyNames": {
              "pattern": "^(?:[a-z][a-z0-9._-]*|user/[A-Za-z0-9._~-]+):[a-z][a-z0-9_]*$"
            },
            "additionalProperties": {
              "type": "object",
              "required": ["model", "properties"],
              "properties": {
                "model": { "type": "string", "minLength": 1 },
                "inferred_at": { "type": "string", "format": "date-time" },
                "durable": {
                  "type": "boolean",
                  "default": false,
                  "description": "When true, a push task MUST NOT replace this entry. Durable entries hold understanding that accumulated rather than being computed from source — a user correction, a pattern an agent noticed across several objects. A task can never set this on its own output: durable means 'cannot be reproduced by re-running', and task output is by definition what re-running produces. Only agent runs and user-driven writes."
                },
                "properties": { "type": "object" },
                "confidence": { "type": "number", "minimum": 0, "maximum": 1 }
              },
              "additionalProperties": false
            }
          }
        },
        "patternProperties": {
          "^x-[a-z0-9-]+$": {
            "description": "Namespaced extensions. Consumers MUST ignore namespaces they don't understand."
          }
        },
        "additionalProperties": false
      },
...
      "Vibe": {
        "title": "Vibe",
        "description": "A dynamic, owned collection of MediaObjects, plus the state that makes it living: its pull configuration and its inferred block. Vibes contain object references, not copies. Vibes carry no source block — they are authored, not ingested; the omission is the ontology.",
        "type": "object",
        "required": ["rnet_schema", "uri", "title", "owner", "objects"],
        "properties": {
          "rnet_schema": {
            "description": "Protocol version this document conforms to. Required, present from commit one.",
            "const": "0.1"
          },
          "uri": {
            "type": "string",
            "pattern": "^rnet://vibe/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "title": { "type": "string", "minLength": 1, "maxLength": 256 },
          "owner": {
            "type": "string",
            "description": "An rnet://id/ URI. Identity issuance is out of protocol; the URI shape is stable and DID-compatible.",
            "pattern": "^rnet://id/[A-Za-z0-9._~-]+$"
          },
          "objects": {
            "type": "array",
            "description": "Ordered, duplicate-free MediaObject URIs — references, not copies. Stores preserve this order across paginated reads.",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^rnet://object/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "pull": {
            "type": "object",
            "description": "How the Vibe acquires new objects.",
            "required": ["enabled"],
            "properties": {
              "enabled": { "type": "boolean" },
              "sources": {
                "type": "array",
                "description": "Source identifiers, implementation-scoped (e.g. \"skill:ofx@^0.3\").",
                "items": { "type": "string", "minLength": 1 }
              },
              "policy": { "enum": ["append_new", "replace", "suggest_only"] },
              "last_pulled_at": { "type": "string", "format": "date-time" }
            },
            "additionalProperties": false
          },
          "grants": { "type": "array", "items": { "$ref": "#/components/schemas/Grant" } },
          "inferred": {
            "type": "object",
            "description": "A map keyed by writer and task: every key is {writer}:{task}. Memory scoped to this record — some entries are task output recomputed from source, others accumulated from corrections and agent observation and cannot be re-derived. A re-run replaces only its own key, and never a durable entry. Consumers MUST treat entries as advisory. The store's summarize task conventionally writes summary and tags.",
            "propertyNames": {
              "pattern": "^(?:[a-z][a-z0-9._-]*|user/[A-Za-z0-9._~-]+):[a-z][a-z0-9_]*$"
            },
            "additionalProperties": {
              "type": "object",
              "required": ["model", "properties"],
              "properties": {
                "model": { "type": "string", "minLength": 1 },
                "inferred_at": { "type": "string", "format": "date-time" },
                "durable": {
                  "type": "boolean",
                  "default": false,
                  "description": "When true, a push task MUST NOT replace this entry. Durable entries hold understanding that accumulated rather than being computed from source — a user correction, a pattern an agent noticed across several objects — and re-running a task cannot reproduce them. Only agent runs and user-driven writes may set this; a task may never mark its own output durable, or refresh stops working."
                },
                "properties": { "type": "object" },
                "confidence": { "type": "number", "minimum": 0, "maximum": 1 }
              },
              "additionalProperties": false
            }
          }
        },
        "patternProperties": {
          "^x-[a-z0-9-]+$": {
            "description": "Namespaced extensions. Consumers MUST ignore namespaces they don't understand."
          }
        },
        "additionalProperties": false
      },
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string", "minLength": 1 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "code": { "type": "string", "minLength": 1 }
        },
        "additionalProperties": true
      }
    }
  }
}